# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DVinyl API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Mobile/native client API for a DVinyl instance, covering authentication,
    collection, wishlist and sharing management, the full item lifecycle, plugin
    metadata, instance administration, and account settings.

    All routes are mounted under `/api/v1`. Requests and responses are JSON
    (`application/json`), except the file-upload endpoints (`multipart/form-data`),
    the share-link QR endpoint (`image/png`) and the import/export endpoints, whose
    exports answer CSV, ZIP or JSON file downloads.

    Authentication is a JWT Bearer access token obtained from `POST /auth/login`.
    Access tokens live 15 minutes (`expiresIn: 900`); use the rotating refresh
    token from `POST /auth/refresh`. Collection-scoped routes are additionally
    gated on the caller's role in the collection named in the path
    (`viewer` < `editor` < `admin`); instance admins pass every collection role.
    Admin routes require `isAdmin` on the account.

    Error envelope: every failed response is `{ "success": false, "error": "..." }`
    with the HTTP status appropriate to the cause. Success responses are
    resource-shaped (e.g. `{ "item": ... }`, `{ "items": [...], "page": ... }`).

    On an instance that has not completed first-run setup, every `/api/v1`
    request returns `503` with `{ "success": false, "error": "Instance not set up yet" }`.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /collections/{id}/settings:
    get:
      tags: [Collections]
      operationId: getCollectionSettings
      summary: Get a collection's settings
      description: |
        Requires `admin` role on the collection. Returns modules, visibility,
        personalisation and plugin settings as one JSON resource. `theme` is a
        `collectionType`/`home` -> preset-key map, and `modules`/`theme` always cover
        every registered plugin.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: The collection's settings.
          content:
            application/json:
              schema:
                type: object
                required: [settings]
                properties:
                  settings: { $ref: '#/components/schemas/CollectionSettings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    patch:
      tags: [Collections]
      operationId: patchCollectionSettings
      summary: Update a collection's settings
      description: |
        Requires `admin` role. Partial update: send only the keys to change. `modules`,
        `theme`, `pluginSettings` and `visibility` shallow-merge; arrays and scalars
        replace. Validation happens before any write, so a rejected request changes
        nothing. Returns the full updated resource.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CollectionSettings' }
      responses:
        '200':
          description: The updated settings.
          content:
            application/json:
              schema:
                type: object
                required: [settings]
                properties:
                  settings: { $ref: '#/components/schemas/CollectionSettings' }
        '400':
          description: Invalid or empty patch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'Unknown module: vinyl' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /collections/{id}/shelves:
    get:
      tags: [Shelves]
      operationId: getCollectionShelves
      summary: List a collection's shelf vocabulary
      description: |
        Requires `viewer` role. Every compartment name in the collection's furniture,
        plus any item `location` that names none of them (so a stray left by an import
        stays pickable), deduplicated and sorted. The same list the confirm route returns
        as `suggestions.location`.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: The shelf names.
          content:
            application/json:
              schema:
                type: object
                required: [shelves]
                properties:
                  shelves:
                    type: array
                    items: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /collections/{id}/furniture:
    get:
      tags: [Shelves]
      operationId: getCollectionFurniture
      summary: List a collection's furniture
      description: |
        Requires `viewer` role. The collection's pieces in `order`, each with its cells
        and, per cell, how many of the collection's owned items stand there (`count`),
        under the caller's visibility filters. A wishlist item and an item of another
        collection are not counted.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: The furniture.
          content:
            application/json:
              schema:
                type: object
                required: [furniture]
                properties:
                  furniture:
                    type: array
                    items: { $ref: '#/components/schemas/Furniture' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      tags: [Shelves]
      operationId: createFurniture
      summary: Create a piece of furniture
      description: |
        Requires `editor` role. Builds an empty piece; cells are added by saving it.
        `layout` must be `cubes` or `rows`; `columns`/`rows` clamp to 1-12; `order`
        defaults after the collection's existing pieces.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FurnitureInput' }
      responses:
        '201':
          description: The new piece.
          content:
            application/json:
              schema:
                type: object
                required: [furniture]
                properties:
                  furniture: { $ref: '#/components/schemas/Furniture' }
        '400':
          description: Empty name or an unknown layout.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'A layout must be cubes or rows' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /collections/{id}/furniture/{furnitureId}:
    put:
      tags: [Shelves]
      operationId: saveFurniture
      summary: Save a piece of furniture whole
      description: |
        Requires `editor` role. Cells arrive in reading order and carry no coordinates;
        the grid is derived from the order and the column count, so a compartment can
        never sit outside its own piece and changing the width reflows it. `from` on a
        cell is the key it had before: a renamed shelf takes its items with it. `409`
        names the clashing shelf (`shelf`) or the piece that already holds it (`furniture`).
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - $ref: '#/components/parameters/FurnitureId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FurnitureInput' }
      responses:
        '200':
          description: The saved piece.
          content:
            application/json:
              schema:
                type: object
                required: [furniture, renamed, moved]
                properties:
                  furniture: { $ref: '#/components/schemas/Furniture' }
                  renamed: { type: integer, description: How many cells were renames. }
                  moved: { type: integer, description: How many items followed them. }
        '400':
          description: Empty name or an unknown layout.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'name is required' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: A shelf name clashes.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'A shelf named "Salon" already exists in this piece of furniture', shelf: Salon }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      tags: [Shelves]
      operationId: deleteFurniture
      summary: Delete a piece of furniture
      description: |
        Requires `editor` role. The items that stood in it keep their `location`, so they
        fall into the reserve and rebuild onto the shelf if it is created again.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - $ref: '#/components/parameters/FurnitureId'
      responses:
        '200':
          description: The piece is gone.
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /collections/{id}/shelf/cell/move:
    post:
      tags: [Shelves]
      operationId: moveShelfCell
      summary: Move a shelf to another piece of furniture
      description: |
        Requires `editor` role. Carries one compartment, with the items that stay standing
        on it untouched, to another piece. `404` names a missing shelf or piece; `moved` is
        false when it is already there.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, to]
              properties:
                key: { type: string, description: The shelf's normalized key. }
                to: { type: string, description: The target piece's id. }
      responses:
        '200':
          description: Whether anything moved.
          content:
            application/json:
              schema:
                type: object
                required: [moved]
                properties:
                  moved: { type: boolean }
        '400':
          description: Missing or invalid key/to.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'key and to are required' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /collections/{id}/shelf/move:
    post:
      tags: [Shelves]
      operationId: moveItemsToShelf
      summary: File items onto a shelf, or take them off
      description: |
        Requires `editor` role. Canonicalises `location` through the shelf store (creating
        the compartment for a new name) and applies it to the collection's own items, up to
        500 at a time. An empty `location` takes them off their shelf. Ids the collection
        does not hold are ignored.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: string }
                location: { type: string, description: The shelf name, or empty to clear it. }
      responses:
        '200':
          description: What moved.
          content:
            application/json:
              schema:
                type: object
                required: [moved, matched, location]
                properties:
                  moved: { type: integer }
                  matched: { type: integer }
                  location: { type: string, description: The canonical spelling applied. }
        '400':
          description: No usable id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'ids must be a non-empty array' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /auth/login` (15-minute lifetime).
  parameters:
    CollectionId:
      name: id
      in: path
      required: true
      description: Collection id (Mongo ObjectId).
      schema: { type: string }
    FurnitureId:
      name: furnitureId
      in: path
      required: true
      description: A piece of furniture's id.
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing, invalid or expired Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing:
              value: { success: false, error: 'Missing bearer token' }
            invalid:
              value: { success: false, error: 'Invalid or expired token' }
    Forbidden:
      description: Authenticated, but lacking the required role or instance-admin flag.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Forbidden' }
    NotFound:
      description: The named resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Collection not found' }
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Internal server error' }
    ServiceUnavailable:
      description: The instance has not completed first-run setup; all `/api/v1` routes return this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Instance not set up yet' }
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
        code:
          type: string
          description: Optional machine-readable code; not used by every handler.
      example: { success: false, error: 'Invalid email or password' }
...
    FurnitureCell:
      type: object
      required: [name, key, row, column]
      properties:
        name: { type: string, description: The compartment's name; items carry it as their `location`. }
        key: { type: string, description: Normalized lookup form, lowercase. }
        row: { type: integer, minimum: 0 }
        column: { type: integer, minimum: 0 }
        capacity: { type: integer, minimum: 0, description: Indicative only; never enforced. }
        count:
          type: integer
          minimum: 0
          description: Owned items standing here, under the caller's visibility filters.
    FurnitureCellInput:
      type: object
      required: [name]
      properties:
        name: { type: string, description: The compartment's name. }
        capacity: { type: integer, minimum: 0, description: Indicative only. }
        from:
          type: string
          description: |
            The cell's previous `key`. Present when this cell is a rename, so the
            items standing on the shelf move with it.
    FurnitureInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 60 }
        layout: { type: string, enum: [cubes, rows] }
        columns: { type: integer, minimum: 1, maximum: 12 }
        rows: { type: integer, minimum: 1, maximum: 12 }
        order: { type: integer, minimum: 0 }
        cells:
          type: array
          items: { $ref: '#/components/schemas/FurnitureCellInput' }
    Furniture:
      type: object
      required: [id, name, layout, columns, rows, order, cells]
      properties:
        id: { type: string }
        name: { type: string }
        layout: { type: string, enum: [cubes, rows] }
        columns: { type: integer, minimum: 1, maximum: 12 }
        rows: { type: integer, minimum: 1, maximum: 12 }
        order: { type: integer }
        cells:
          type: array
          items: { $ref: '#/components/schemas/FurnitureCell' }
    CollectionSettings:
      type: object
      properties:
        modules:
          type: object
          description: One entry per registered plugin's collectionType.
          additionalProperties: { type: boolean }
        mergeDuplicates: { type: boolean }
        pluginSettings:
          type: object
          additionalProperties:
            type: object
            additionalProperties: true
        visibility:
          type: object
          properties:
            applyToAdmin: { type: boolean }
            hiddenItems:
              type: array
              items: { type: string }
            hiddenGenres:
              type: array
              items: { type: string }
            hiddenTypes:
              type: array
              description: Plugin `kind`s (e.g. `Music`), not collectionTypes.
              items: { type: string }
        theme:
          type: object
          description: '`home` or a collectionType, mapped to a preset key.'
          additionalProperties: { type: string }
        aspectRatioClass: { type: string }
        navbarShortcuts:
          type: array
          items: { type: string }
        statsWidgets:
          type: array
          items: { type: string }
        fastAdd: { type: string }
...